projects/01-Fraud-Interception-API/requirements.txt

azure-functions
xgboost>=2.0.0
scikit-learn
pandas
numpy
imbalanced-learn
PyYAML
joblib